and now...
I wanted to really get people to
the classest video on the planet.

Kenny G is playing a sax.

He looks a fool but prolly has lots of
beautiful dudes wanting that circular
breathing thing he can do.

The funny thing is that if he had a
diggeree-doo instead of a sop-sax
he'd be 1000x cooler? think about it.